Update building_block_integration.md
Fixed grammar. Sentence doesn't make sense with "This" followed directly the name. See example for "Scenes" integration below.

CURRENT:
This scenes is a building block integration that cannot be added to your Home Assistant directly but is used and provided by other integrations.

PROPOSED:
Scenes is a building block integration that cannot be added to your Home Assistant directly but is used and provided by other integrations.

OR:
This is a building block integration that cannot be added to your Home Assistant directly but is used and provided by other integrations.